Clear "redefinition of class UC1701" error
Depending on the program structure, header files can be included more than once in the program code. This means that the class would be redefined, which results in a compiler error. To avoid this, I added this standard macro to make the preprocessor add the code only once and therefore avoid the said compiler error.

Compiler error:
"error: redefinition of 'class UC1701'
 class UC1701: public Print {"